updated "known result" for polariton_energy for new fields::integrate-based energy The new field_energy function, using fields::integrate, gives slightly different energy results from the old code.  This is because of different integration weights for the boundary edges...I think the new version is more reasonable, actually.
Basically, the old code just summed up the energy for every point in
the cell, with uniform integration weight.  The new code is almost the
same, except that at the boundaries, it integrates the boundary pixels
assuming linear interpolation to the points just beyond (or at) the
boundary.  (Points beyond the boundary are assumed zero for
non-periodic boundary conditions.)  I think this should be more
accurate.

Of course, the new result and the old result converge to one another
as the resolution is increased.