make sure coordinates are exact in symmetry.cpp In symmetry.cpp, we want to check that adding symmetry to a simulation produces *exactly* the same result as the simulation without explicit use of symmetry.  This, however, requires a great deal of care with floating point arithmetic.
In particular, one important requirement is that the sources *exactly*
obey the symmetry condition.  In order for this to be true, the
location of the source *as represented within floating point* must be
symmetric.  (Recall that floating-point can only represent base-2
rational fractions exactly, i.e. integer * 2^e.)  To impose this
condition, we employ two tricks:

First, the resolution 'a' should be exactly representable (e.g. a
power of 2) , so that no surprises occur when converting vecs to ivecs
etc.

Second, the coordinates, or at least those coordinates which are
supposed to be symmetric, must be exactly representable.

For example, suppose that something is supposed to be mirror symmetric
around x=0.5.  You might think that sources at x=0.2 and x=0.8 would
satisfy this, but they don't because those two coordinates are not
exactly representable--they are stored as 0.2+eps1 and 0.8+eps2, where
eps1 and eps2 are ~1e-16, and these coordinates are not symmetric.
Instead, we use e.g. x=0.25 and x=0.75, which are symmetric because
they are exactly represented (1 / 2^2 and 3 / 2^2, respectively).

Of course, a user code is not expected to play these games, but in
such circumstances it is acceptable for the symmetric and
non-symmetric simulations to differ by errors on the order of the
machine precision.