rename c -> Courant, make a field of structure instead of global First of all, I renamed "c" to "Courant", since that is what it is (Courant number = dt/dx) ... it is not the speed of light, which should be 1.  (I've run into bugs multiple times now from having "c" be shadowed by a local variable, e.g. for the component, and this gets rid of that forever.)
Second, since people may want to change the Courant number to reduce
numerical dispersion (in general, it only has to be < 1 /
sqrt(#dimensions)), I made it a field of the structure class (and also
of structure_chunk and fields_chunk).  The default is still 0.5,
however.

Third, inva is no longer a field of the fields class; instead, we have
a new field dt which is the timestep.  This is much more intuitive and
useful to the end-user (and to us) instead of having to compute
Courant*inva all the time.

There some minor related changes.  e.g. do_harminv takes simply the
sampling interval dt instead of a scale_factor and a.