completely revamped structure and structure_chunk initialization In the previous version of the code, the typical initialization sequence went like this:
1) create structure (initializes chunks, evaluates epsilon)
2) set PML (re-initializes chunks, shuffles epsilon around)

Step (2) is problematic because the communication of epsilon to the
new chunks was performed point-by-point.  Not only is this incredibly
slow, but it actually causes Tungsten's MPI implementation to crash
(probably because there are too many messages).  Moreover, it is
completely unnecessary.

In the new version, the PML boundaries are specified when the structure
is first created, so that the chunks need only be computed once and are
never re-shuffled around.

INCOMPATIBLE API CHANGE:

  old:
     structure s(vol, eps, splitting, symmetries);
     s.use_pml_everywhere(1.0);

  new:
     structure s(vol, eps, pml(1.0), symmetries, splitting);

In general, the PML boundary regions are specified by *summing* terms
of the form pml(thickness, direction, High/Low).  If one or both of
the last two arguments is omitted, it means *all* sides and *all*
directions, respectively.  So, a more complicated example is:

  old:
     structure s(vol, eps, splitting, symmetries);
     s.use_pml(X, High, 1.0);
     s.use_pml(X, Low, 1.0);
     s.use_pml(Y, High, 1.0);

  new:
     structure s(vol, eps, pml(1.0, X) + pml(1.0, Y, High),
                 symmetries, splitting);

To specify *no* PML, you use:

   structure s(vol, eps, no_pml(), symmetries, splitting);

This is actually the default, if you omit the boundary-region argument
along with splitting (default = 0) and symmetries (default =
identity).

INCOMPATIBLE API CHANGE:

  old:
     s.set_epsilon(eps, minvol, do_aniso_averaging);

  new:
     s.set_epsilon(eps, do_aniso_averaging, minvol);

The default for the last two arguments is true and 0, respectively, so
in the usual case where you are using this function because you want
to turn on averaging (currently not the default), you just do:

   s.set_epsilon(eps);