major surgery on connect_the_chunks connect_the_chunks was:    (a) a major performance bottleneck on multiprocessor systems    (b) very difficult to modify, consisting of the same code        copy-and-pasted many times    (c) fairly fragile, with the loops impossible to reorder        without breaking step.cc
I've tried to fix all of those problems.  Now, instead of calling
things like locate_point O(nc^2) where np is the number of chunks,
it only computes it O(nc) times, and a lot of the duplication is
removed.  It's also much less fragile, because the connections are
stored per phase type (complex, -1, or +1) rather than in one
order-sensitive array.

However, this required major changes to boundaries.cpp and related code.