add nonlinear symmetry tests
The nonlinear symmetry tests had been disabled because they caused
the symmmetry.dac test to fail.  I had assumed this was a bug,
but now I'm not so sure.

First, the nonlinear calculations rely on an averaging of field values
from adjacent points, and this averaging process will differ by rounding
errors when using symmetries (e.g. left + right + top + bottom may
change to right + left + top + bottom, which differs by rounding).  So,
I had to decrease the error threshold for the nonlinear tests.

Second, a more serious concern: symmetry is not necessarily preserved
when rounding is included.  A simple example is that exp(i m phi) modes
in a cylindrical waveguide may be converted to exp(3 i m phi) modes
by a Kerr nonlinearity, analogous to 3rd harmonic generation.  I *think*
that this conversion does not occur for the trivial representation
(analogous to m=0), but I'm not 100% sure.  So, the nonlinear tests
(which use the trivial representation, i.e. no phase factor) should
be okay...

However, e.g. the rot4z symmetry test still fails if I use *both* TM
and TE sources (which interact due to the nonlinearity).  Outputting
the fields, it looks like a nonsymmetric pattern really is generated
after a while in this case.   Perhaps this is a numerical effect,
related to the rounding mentioned above?