Minor bug fixes to 3 `ModeSolver` methods (#1125)
Just fixing the return values of `get_poynting`. Closes #1120. Also update `get_charge_density` to be consistent with method signature currently in the docs. 
Last, `compute_field_divergence` seemed to break the pattern of of its neighboring methods. Bug that it did not make a call to `self` fixed, and I assume that we also expect it to return the field divergence after it has been computed.